+++ eslint-plugin-expensify/prefer-react-class-constructors-with-props.js
